コメント機能の修正・mainの機能分割
Before: main.py (全機能が1ファイル)
After:
├── app.py                 # アプリケーションエントリーポイント
├── config.py              # 設定・定数定義
├── utils/
│   ├── json_utils.py      # JSON操作ユーティリティ
│   └── file_utils.py      # ファイル操作ユーティリティ
└── routes/
    ├── auth.py            # 認証機能
    ├── posts.py           # 投稿機能
    ├── main.py            # メインページ
    └── api.py             # API機能
