+++ Regions.json
@@ -16,5 +16,8 @@
   },
   "e2a47a1c-f246-4eba-b12e-f656edaff393": {
     "region": "沖縄"
+  },
+  "39a04753-b6ea-447d-b14a-9b387cf15a9c": {
+    "region": "東海圏"
   }
 }
