+++ Tags.json
@@ -38,5 +38,13 @@
     "映え",
     "料理",
     "スポーツ"
+  ],
+  "39a04753-b6ea-447d-b14a-9b387cf15a9c": [
+    "景色",
+    "動物",
+    "スイーツ",
+    "映え",
+    "料理",
+    "スポーツ"
   ]
 }
