+++ templates/login.html
@@ -16,10 +16,11 @@
         <input type="password" id="password" name="password" required>
     </div>
 
-    <button type="submit" class="btn">決定</button>
+    <button type="submit" class="btn" style="min-width: 150px; width: 150px;">決定</button>
 </form>
 
 <div style="margin-top: 20px;">
-    <a href="{{ url_for('register') }}" class="btn btn-secondary">ユーザー登録</a>
+    <a href="{{ url_for('auth.register') }}" class="btn btn-secondary"
+        style="min-width: 150px; width: 150px; text-align: center;">ユーザー登録</a>
 </div>
 {% endblock %}
