+++ templates/map.html
@@ -16,7 +16,7 @@
 <body>
   <h1>{{ region }}の投稿マップ</h1>
   <div id="map"></div>
-  <a href="{{ url_for('home') }}">← ホームに戻る</a>
+  <a href="{{ url_for('main.home') }}">← ホームに戻る</a>
 
   <script src="https://unpkg.com/leaflet@1.9.4/dist/leaflet.js"></script>
   <script>
