+++ templates/post.html
@@ -47,8 +47,11 @@
         </select>
     </div>
 
-    <button type="submit" class="btn">投稿</button>
-    <a href="{{ url_for('home') }}" class="btn btn-secondary">キャンセル</a>
+    <div style="display: flex; gap: 15px; margin-top: 20px;">
+        <button type="submit" class="btn" style="min-width: 120px;">投稿</button>
+        <a href="{{ url_for('main.home') }}" class="btn btn-secondary"
+            style="min-width: 120px; text-align: center;">キャンセル</a>
+    </div>
 </form>
 
 <script>
