+++ templates/register.html
@@ -24,10 +24,13 @@
         <input type="password" id="password" name="password" required>
     </div>
 
-    <button type="submit" class="btn">登録</button>
-</form>
+    <div style="display: flex; gap: 15px; margin-top: 20px;">
+        <button type="submit" class="btn" style="min-width: 120px;">登録</button>
+    </div>
 
-<div style="margin-top: 20px;">
-    <a href="{{ url_for('login') }}" class="btn btn-secondary">戻る</a>
-</div>
+    <div style="margin-top: 20px;">
+        <a href="{{ url_for('auth.login') }}" class="btn btn-secondary"
+            style="min-width: 120px; text-align: center;">戻る</a>
+    </div>
+</form>
 {% endblock %}
